first attempt: back button returns to review step
not working well. i refactored the triggerreveveiw function so that the review button listener is not part of the function. because otherwise it craps out. you can't trigger a function by clicking the back button when the function expects to be triggered when the review button is clicked.

I also started working on actually saving the review table's configurations. that way, we remember the config when the user comes back - and will be useful for analyze.