Upload directly to Analyze
I moved the entire Review step into a settings menu of Analyze. the 'guess' function i created previously makes it so that the review mapping is accurate >80% of the time, eliminating the need to tax the user on learning how to map their data. one less step for the user