Brought back the review step
way of showing the data I uploaded (back to the review step) ‘does this look OK? ‘ we took the liberty to categorize but may not be right. You can also change this later in  data settings
Need a retry button - brings back to upload
Relayout
Missing some functions - it was that the parsecsv function was taking too long and the genReviewTable function was getting impatient and ran without it having completed
If i leave the tool from here, and come back, i should be navigated back here, not analysis step
Now as sson as i go to analyze step, whatever changes i did in review are lost because i guess data types as soon as i get to analysis step.
First problem is that review step is useless since ilose changes
Second is that this was never good even prior to review step reincorporation because i would lose any changes i did in a previous session. I never stored dropdownstate